Deprecate PhantomJS based report. Introduce Headless Chrome based report.
