+++ package.json
@@ -37,6 +37,7 @@
     "mustache": "2.3.0",
     "node-horseman": "3.3.0",
     "node-slack": "0.0.7",
+    "puppeteer": "1.2.0",
     "sum-time": "1.0.0",
     "uuid": "3.1.0"
   },
