+++ public/constants/report_watcher.js
@@ -20,7 +20,6 @@ export default {
         snapshot: {
           res: '1280x900',
           url: 'http://localhost/app/kibana#/dashboard/Alerts',
-          path: '/tmp/',
           params: {
             delay: 5000,
             crop: 'false'
