+++ public/messages/help.json
@@ -2,7 +2,7 @@
   "schedule": "<p> <i>fires at 17:15 every day </i> <br/> <code> at 17:15 </code> </p> <p> <i>fires every 5 minutes every day </i> <br/> <code> every 5 mins </code> </p> <p> <i>fires at 10:15am and 5:15pm every day except on Tuesday </i> <br/> <code> at 10:15 am also at 5:15pm except on Tuesday </code> </p> <p> <i>fires every Sunday </i> <br/> <code> on the first day of the week </code> </p> <p> <i>fires on the last day of every month </i> <br/> <code> on the last day of the month </code> </p> <p> <i>fires every day between the 15th and 20th day every month </i> <br/> <code> on the 15th through 20th day of the month </code> </p> <p> <i>fires every 5 minutes on Saturday and Sunday </i> <br/> <code> every 5 mins every weekend </code> </p> <p> <i>fires on the 7th, 27th, and 47th minute every hour </i> <br/> <code> every 20 mins starting on the 7th min </code> </p> <p> <i>fires every hour after noon every day </i> <br/> <code> after 12th hour </code> </p> <p> <i>fires every hour before noon every day </i> <br/> <code> before 12th hour </code> </p> <p> <i>fires at 5:00pm every day </i> <br/> <code> at 5:00 pm </code> </p> <p> <i>fires at 5:00pm on Wednesday, Thursday, and Friday </i> <br/> <code> at 5:00 pm on Weds,Thurs and Fri </code> </p> <p> <i>fires at 5:00pm every day of March in 2014 </i> <br/> <code> at 5:00 pm every 1 day of March in 2014 </code> </p>",
   "action": {
     "report": {
-      "path": "Screenshot filename will be generated automatically. You can specify the filename, for example, myreport.png."
+      "path": "Screenshot filename will be generated automatically. You can specify the filename, for example, my_report"
     },
     "save_payload": "Store all Elasticsearch documents found by Input query inside alarm document. Attention, this can freeze your browser if you store a lot of documents!"
   }
