+++ server/lib/actions/helpers/email.js
@@ -4,40 +4,27 @@ import Promise from 'bluebird';
 /**
 * Email server
 */
-export default class Email {
+class Email {
 
   /**
   * Instantiate email server object
   *
-  * @param {object} args - init server connection settings (user, password, host, ssl, timeout)
+  * @param {object} options to init server connection settings (user, password, host, ssl, timeout)
   */
-  constructor(args) {
-    const { user, password, host, ssl, timeout } = args;
-    this.server = emailjs.server.connect({
-      user,
-      password,
-      host,
-      ssl,
-      timeout
-    });
+  constructor(options) {
+    this.server = emailjs.server.connect(options);
   }
 
   /**
   * Send email
   *
-  * @param {object} args - email parameters (text, from, to, subject, attachment)
+  * @param {object} options to send email (text, from, to, subject, attachment)
   */
-  send(args) {
-    const { text, from, to, subject, attachment } = args;
+  send(options) {
     return new Promise((resolve, reject) => {
-      this.server.send({
-        text,
-        from,
-        to,
-        subject,
-        attachment
-      }, function (error, message) {
+      this.server.send(options, function (error, message) {
         if (error) {
+          console.error('fail to send email', error);
           reject(error);
         }
         resolve(JSON.stringify(message));
@@ -45,3 +32,5 @@ export default class Email {
     });
   }
 }
+
+export default Email;
