+++ server/lib/actions/index.js
@@ -1,0 +1,2 @@
+import doActions from './do_actions';
+export default doActions;
