+++ server/lib/classes/watcher.js
@@ -6,7 +6,7 @@ import compare from '../validators/compare';
 import compareArray from '../validators/compare_array';
 import getElasticsearchClient from '../get_elasticsearch_client';
 import getConfiguration from '../get_configuration';
-import actionFactory from '../actions/actions';
+import actionFactory from '../actions';
 import Log from '../log';
 import { isKibi } from '../helpers';
